+++ dailimage.go
@@ -1,21 +1,41 @@
 package main
 
 import (
+	"embed"
 	"fmt"
-	"local/eldritchgarden/dailimage/config"
-	"local/eldritchgarden/dailimage/image"
+	"html/template"
+	"io/fs"
 	"net/http"
 
+	"github.com/eldritchgarden/dailimage/image"
+	"github.com/eldritchgarden/dailimage/internal/config"
+
 	"github.com/gin-gonic/gin"
+	log "github.com/sirupsen/logrus"
 )
+
+//go:embed templates
+var templateFS embed.FS
+
+//go:embed static
+var staticFS embed.FS
 
 func main() {
 	router := gin.Default()
 	if len(config.ENV.TrustedProxies) > 0 {
 		router.SetTrustedProxies(config.ENV.TrustedProxies)
+		log.Debug("Trusted proxies: ", config.ENV.TrustedProxies)
 	} else {
 		router.SetTrustedProxies(nil) // Disable trusted proxies if none are set
+		log.Debug("Trusted proxies disabled")
 	}
+
+	// Load embedded templates
+	router.SetHTMLTemplate(template.Must(template.New("").ParseFS(templateFS,
+		"templates/*")))
+
+	staticFS, _ := fs.Sub(staticFS, "static")
+	router.StaticFS("/static", http.FS(staticFS))
 
 	router.GET("/ping", func(c *gin.Context) {
 		c.JSON(http.StatusOK, gin.H{
@@ -25,6 +45,7 @@ func main() {
 
 	router.GET("/random", image.GetRandomImage)
 	router.GET("/random/*subpath", image.GetRandomImagePath)
+	router.GET("/slideshow", image.GetSlideshow)
 
 	router.Run(fmt.Sprintf("%s:%s", config.ENV.Addr, config.ENV.Port))
 }