+++ changelog.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## 0.1.1 (unreleased)
+## 0.2.0 (unreleased)
 - Added example scripts for changing wallpaper on Windows and MacOS
+- Added slideshow page
+- Improved image selection with variable uniqueness to prevent repeats
+- Fixed headers to prevent caching images
+- Release binaries for linux and mac
+- Release arm64 binaries/images
+- Restructured repo for Go standards
 
 ## 0.1.0
 First release